Move tests for new donations
Some tests in AddDonationRouteTest were actually testing the "new
donation" route.
