+++ tests/EdgeToEdge/Routes/AddDonationRouteTest.php
@@ -657,70 +657,6 @@ class AddDonationRouteTest extends WebRouteTestCase {
 		} );
 	}
 
-	public function testWhenTrackableInputDataIsSubmitted_theyAreStoredInSession(): void {
-		$this->createAppEnvironment( [], function ( Client $client, FunFunFactory $factory, Application $app ): void {
-			$client->request(
-				'GET',
-				'/',
-				[
-					'betrag' => '5,00',
-					'periode' => 3,
-					'zahlweise' => 'BEZ'
-				]
-			);
-
-			$piwikTracking = $app['session']->get( 'piwikTracking' );
-			$this->assertSame( 'BEZ', $piwikTracking['paymentType'] );
-			$this->assertSame( 3, $piwikTracking['paymentInterval'] );
-			$this->assertSame( '5,00', $piwikTracking['paymentAmount'] );
-		} );
-	}
-
-	public function testWhenTolstojNovelIsPassed_isIsNotStoredInSession(): void {
-		$this->createAppEnvironment( [], function ( Client $client, FunFunFactory $factory, Application $app ): void {
-
-			$client->request(
-				'GET',
-				'/',
-				[
-					'betrag' => '5,00',
-					'periode' => 3,
-					'zahlweise' => 'Eh bien, mon prince. Gênes et Lucques ne sont plus que des apanages, des поместья, de la ' .
-						'famille Buonaparte. Non, je vous préviens que si vous ne me dites pas que nous avons la guerre, si ' .
-						'vous vous permettez encore de pallier toutes les infamies, toutes les atrocités de cet Antichrist ' .
-						'(ma parole, j’y crois) — je ne vous connais plus, vous n’êtes plus mon ami, vous n’êtes plus мой ' .
-						'верный раб, comme vous dites. Ну, здравствуйте,' .
-						'здравствуйте. Je vois que je vous fais peur, ' .
-						'садитесь и рассказывайте.'
-				]
-			);
-
-			$piwikTracking = $app['session']->get( 'piwikTracking' );
-			$this->assertArrayNotHasKey( 'paymentType', $piwikTracking );
-			$this->assertSame( 3, $piwikTracking['paymentInterval'] );
-			$this->assertSame( '5,00', $piwikTracking['paymentAmount'] );
-		} );
-	}
-
-	public function testWhenParameterIsOmitted_itIsNotStoredInSession(): void {
-		$this->createAppEnvironment( [], function ( Client $client, FunFunFactory $factory, Application $app ): void {
-
-			$client->request(
-				'GET',
-				'/',
-				[
-					'betrag' => '5,00',
-					'zahlweise' => 'BEZ'
-				]
-			);
-
-			$piwikTracking = $app['session']->get( 'piwikTracking' );
-			$this->assertSame( 'BEZ', $piwikTracking['paymentType'] );
-			$this->assertSame( '5,00', $piwikTracking['paymentAmount'] );
-			$this->assertArrayNotHasKey( 'paymentInterval', $piwikTracking );
-		} );
-	}
-
 	public function testWhenInitiallyIntendedPaymentOptionsDifferFromActual_itIsReflectedInPiwikTrackingEvents(): void {
 		$this->createEnvironment( [], function ( Client $client, FunFunFactory $factory ): void {
 			$factory->setEmailValidator( new EmailValidator( new NullDomainNameValidator() ) );
@@ -854,33 +790,6 @@ class AddDonationRouteTest extends WebRouteTestCase {
 			$this->assertTrue( $cookie->isSecure() );
 			$this->assertTrue( $cookie->isHttpOnly() );
 		} );
-	}
-
-	public function testAllPaymentTypesAreOffered(): void {
-		$client = $this->createClient( [ 'skin' => [ 'default' => '10h16' ] ] );
-		$client->request(
-			'GET',
-			'/donation/new'
-		);
-		$crawler = $client->getCrawler();
-
-		$this->assertSame( 1, $crawler->filter( '#donation-payment input[name="zahlweise"][value="BEZ"]' )->count() );
-		$this->assertSame( 1, $crawler->filter( '#donation-payment input[name="zahlweise"][value="UEB"]' )->count() );
-		$this->assertSame( 1, $crawler->filter( '#donation-payment input[name="zahlweise"][value="MCP"]' )->count() );
-		$this->assertSame( 1, $crawler->filter( '#donation-payment input[name="zahlweise"][value="PPL"]' )->count() );
-		$this->assertSame( 1, $crawler->filter( '#donation-payment input[name="zahlweise"][value="SUB"]' )->count() );
-	}
-
-	public function testSofortPaymentTypeCanByDisabledViaQuery(): void {
-		$client = $this->createClient( [ 'skin' => [ 'default' => '10h16' ] ] );
-		$client->request(
-			'GET',
-			'/donation/new',
-			[ 'pmt' => '0' ]
-		);
-		$crawler = $client->getCrawler();
-
-		$this->assertSame( 0, $crawler->filter( '#donation-payment input[name="zahlweise"][value="SUB"]' )->count() );
 	}
 
 	public function testDonationReceiptOptOut_persistedInDonation(): void {
